requirements.txt
langchain
langchain_openai
python-dotenv
langchain_anthropic
langchain_google_genai
langchain_huggingface
sentence-transformers
pydantic
email-validator
grandalf